Move mapping logic to a helper class
